# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ActivityDay {
  day: string;
  minutes: number;
  isToday?: boolean;
}

export interface Dashboard {
  learnerName: string;
  greeting: string;
  currentTopic: string;
  currentTopicLabel?: string;
  progressPercent: number;
  streakDays: number;
  weeklyMinutes: number;
  completedTopics: number;
  totalTopics: number;
  nextAction: string;
  weeklyActivity: ActivityDay[];
  focusAreas?: string[];
}

export type LearningPathNodeStatus = typeof LearningPathNodeStatus[keyof typeof LearningPathNodeStatus];


export const LearningPathNodeStatus = {
  completed: 'completed',
  current: 'current',
  upcoming: 'upcoming',
  locked: 'locked',
} as const;

export interface LearningPathNode {
  id: string;
  title: string;
  category: string;
  description: string;
  status: LearningPathNodeStatus;
  durationMinutes: number;
  order: number;
  accent?: string;
  confidence?: number;
}

export interface Topic {
  id: string;
  title: string;
  category: string;
  description: string;
  lessonCount: number;
  color: string;
  progressPercent?: number;
}

export interface PracticePrompt {
  id: string;
  topic: string;
  title: string;
  prompt: string;
  options: string[];
  minutes: number;
  difficulty: string;
}

export interface DiagnosticInput {
  topic: string;
  answer: string;
  claimedLevel: string;
}

export interface DiagnosticResult {
  evaluatedLevel: string;
  accuracyPercent: number;
  strengths: string[];
  focusAreas: string[];
  message: string;
}

export interface OnboardingInput {
  userName: string;
  gender: string;
  ageGroup: string;
  role: string;
  focus: string;
  answers: string[];
}

export interface OnboardingResult {
  evaluatedLevel: string;
  accuracyPercent: number;
  focusLabel: string;
  pathSummary: string;
  message: string;
}

export interface PracticeCompletionInput {
  promptId: string;
  answer: string;
}

export interface PracticeResult {
  correct: boolean;
  scorePercent: number;
  feedback: string;
  nextStep: string;
}

export interface Peer {
  id: string;
  name: string;
  role: string;
  focus: string;
  level: string;
  matchPercent: number;
  initials: string;
  accent: string;
}

export interface TutorMessageInput {
  message: string;
  /** @nullable */
  topic?: string | null;
}

export interface TutorMessage {
  reply: string;
  topic: string;
  suggestedPrompt: string;
}