# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /diagnostic/submit:
    post:
      operationId: submitDiagnostic
      tags: [learning]
      summary: Submit an answer from the adaptive diagnostic
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DiagnosticInput"
      responses:
        "200":
          description: Updated learner evaluation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DiagnosticResult"
  /onboarding/complete:
    post:
      operationId: completeOnboarding
      tags: [learning]
      summary: Complete learner onboarding and generate a path
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OnboardingInput"
      responses:
        "200":
          description: Personalized learner path summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OnboardingResult"
...
  /practice/complete:
    post:
      operationId: completePractice
      tags: [learning]
      summary: Complete the daily practice prompt
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PracticeCompletionInput"
      responses:
        "200":
          description: Practice result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PracticeResult"
...
  /tutor/message:
    post:
      operationId: sendTutorMessage
      tags: [learning]
      summary: Ask the EduAgent tutor a question
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TutorMessageInput"
      responses:
        "200":
          description: Tutor reply
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TutorMessage"
...
components:
  schemas:
...
    DiagnosticInput:
      type: object
      required: [topic, answer, claimedLevel]
      properties:
        topic: { type: string }
        answer: { type: string }
        claimedLevel: { type: string }
    DiagnosticResult:
      type: object
      required: [evaluatedLevel, accuracyPercent, strengths, focusAreas, message]
      properties:
        evaluatedLevel: { type: string }
        accuracyPercent: { type: integer }
        strengths:
          type: array
          items: { type: string }
        focusAreas:
          type: array
          items: { type: string }
        message: { type: string }
    OnboardingInput:
      type: object
      required: [userName, gender, ageGroup, role, focus, answers]
      properties:
        userName: { type: string }
        gender: { type: string }
        ageGroup: { type: string }
        role: { type: string }
        focus: { type: string }
        answers:
          type: array
          items: { type: string }
    OnboardingResult:
      type: object
      required: [evaluatedLevel, accuracyPercent, focusLabel, pathSummary, message]
      properties:
        evaluatedLevel: { type: string }
        accuracyPercent: { type: integer }
        focusLabel: { type: string }
        pathSummary: { type: string }
        message: { type: string }
    PracticeCompletionInput:
      type: object
      required: [promptId, answer]
      properties:
        promptId: { type: string }
        answer: { type: string }
    PracticeResult:
      type: object
      required: [correct, scorePercent, feedback, nextStep]
      properties:
        correct: { type: boolean }
        scorePercent: { type: integer }
        feedback: { type: string }
        nextStep: { type: string }
...
    TutorMessageInput:
      type: object
      required: [message]
      properties:
        message: { type: string }
        topic: { type: ["string", "null"] }
    TutorMessage:
      type: object
      required: [reply, topic, suggestedPrompt]
      properties:
        reply: { type: string }
        topic: { type: string }
        suggestedPrompt: { type: string }
        route: { type: string, enum: [graph_rag, standard_rag] }
        citations:
          type: array
          items:
            type: object
            properties:
              label: { type: string }
              relation: { type: string }
...