# lib/api-zod/src/generated/types/diagnosticResult.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface DiagnosticResult {
  evaluatedLevel: string;
  accuracyPercent: number;
  strengths: string[];
  focusAreas: string[];
  message: string;
}