# lib/api-zod/src/generated/types/onboardingInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface OnboardingInput {
  userName: string;
  gender: string;
  ageGroup: string;
  role: string;
  focus: string;
  answers: string[];
}