# lib/api-zod/src/generated/types/onboardingResult.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface OnboardingResult {
  evaluatedLevel: string;
  accuracyPercent: number;
  focusLabel: string;
  pathSummary: string;
  message: string;
}