# lib/api-zod/src/generated/types/practiceCompletionInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface PracticeCompletionInput {
  promptId: string;
  answer: string;
}